requirements.txt
fastapi[standard]>=0.121.0
uvicorn>=0.38.0
llama-cpp-python>=0.3.16
pyyaml>=6.0
